Fixed circular reference, still has issues though
