+++ src/backend/models/GameModel.ts
@@ -1,6 +1,7 @@
 import { ChildEntity, Column, CreateDateColumn, Entity, JoinColumn, JoinTable, ManyToMany, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
 // Instead of direct import, we'll use a type-only import and function references
-import { UserModel } from "./UserModel.js";
+import { UserModel } from "./common.js";
+import {MatchModel} from "./common.js"
 
 @ChildEntity("games")
 export class GameModel extends MatchModel {
@@ -15,7 +16,3 @@ export class GameModel extends MatchModel {
     @Column({ nullable: true })
     gameAdminId?: number;
 }
-
-// Import at the end to avoid the circular dependency issue
-import "./UserModel.js";
-import {MatchModel} from "./MatchModel.js"
