+++ src/backend/models/common.ts
@@ -1,0 +1,5 @@
+
+export * from './UserModel.js'
+export * from './MatchModel.js'
+export * from './GameModel.js'
+export * from './TournamentModel.js'
